+++ gulpfile.js/tasks/build-sass.js
@@ -1,0 +1,28 @@
+const { src, dest } = require('gulp');
+const { outDir } = require('../helpers');
+
+const sass = require('gulp-sass');
+const rename = require('gulp-rename');
+const postcss = require('gulp-postcss');
+const cssnano = require('cssnano');
+sass.compiler = require('sass');
+
+const buildSass = () => {
+  return src('./src/index.scss')
+    .pipe(
+      sass({
+        importer: [
+          require('../../glob-sass-importer'),
+          // Import Theo design tokens as SCSS variables
+          require('../../.theo/sass-importer'),
+        ],
+      }).on('error', sass.logError)
+    )
+    .pipe(rename({ basename: 'standalone' }))
+    .pipe(dest(outDir))
+    .pipe(postcss([cssnano()]))
+    .pipe(rename({ extname: '.min.css' }))
+    .pipe(dest(outDir));
+};
+
+module.exports = buildSass;
