+++ package.json
@@ -59,6 +59,7 @@
     "css-loader": "3.5.3",
     "cssnano": "4.1.10",
     "dts-bundle-generator": "4.3.0",
+    "eliminate": "1.1.2",
     "eslint": "6.8.0",
     "eslint-config-prettier": "6.11.0",
     "eslint-plugin-mdx": "1.6.9",
@@ -80,7 +81,6 @@
     "prettier": "2.0.5",
     "remark-preset-lint-recommended": "4.0.0",
     "remark-preset-prettier": "0.4.0",
-    "require-dir": "1.2.0",
     "rollup": "2.7.6",
     "rollup-plugin-terser": "5.3.0",
     "sass": "1.26.5",
@@ -103,15 +103,15 @@
     "start-storybook": "start-storybook -p 6006 -s static",
     "build-storybook": "npm run preprocess && build-storybook -s static",
     "preprocess": "run-p preprocess:*",
-    "preprocess:svg-to-twig": "gulp svg-to-twig",
-    "preprocess:theo-to-mdx": "gulp theo-to-mdx",
-    "watch-preprocess": "gulp watch-preprocess",
+    "preprocess:svg-to-twig": "gulp svgToTwig",
+    "preprocess:theo-to-mdx": "gulp theoToMDX",
+    "watch-preprocess": "gulp watchPreprocess",
     "build": "gulp build",
     "type": "tsc",
     "type:watch": "tsc --watch",
     "check-lint": "run-s check-lint:*",
     "check-lint:prettier": "prettier . --check",
-    "check-lint:css": "stylelint '**/*.scss'",
+    "check-lint:css": "stylelint '**/-*.scss'",
     "check-lint:js": "eslint . --ext .js,.mdx,.ts,.tsx",
     "lint": "run-s lint:*",
     "lint:prettier": "prettier . --write",
